+++ tests/operations/DeleteTest.cs
@@ -1,9 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Diagnostics;
+using System.IO;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using werkbank.operations;
+using werkbank.services;
 
 namespace tests.operations
 {
@@ -30,6 +33,74 @@ namespace tests.operations
             Delete.Perform(hi);
 
             Assert.IsFalse(File.Exists(hi));
+        }
+
+        [TestMethod]
+        public void PerformWorksWithSymbolicDirLink()
+        {
+            if (!WinApiService.IsUserAnAdmin())
+            {
+                Assert.Inconclusive("This test only works on administrative privileges.");
+                return;
+            }
+
+            string source = Util.GetTempPath();
+            string dest = Directory.CreateDirectory(Util.GetTempPath()).FullName;
+
+            WinApiService.CreateSymbolicLink(source, dest, WinApiService.SYMBOLIC_LINK_FLAG.Directory);
+
+            // delete dest to break the symlink
+            Directory.Delete(dest);
+
+            Delete.Perform(source);
+
+            Assert.IsFalse(Directory.Exists(source));
+        }
+
+        [TestMethod]
+        public void PerformWorksWithNestedSymbolicDirLink()
+        {
+            if (!WinApiService.IsUserAnAdmin())
+            {
+                Assert.Inconclusive("This test only works on administrative privileges.");
+                return;
+            }
+
+            string source = Util.GetTempPath();
+            string subDir = Directory.CreateDirectory(Path.Combine(source, "xoxo", "hi")).FullName;
+            string link = Path.Combine(subDir, "ho");
+            string dest = Directory.CreateDirectory(Util.GetTempPath()).FullName;
+
+            WinApiService.CreateSymbolicLink(link, dest, WinApiService.SYMBOLIC_LINK_FLAG.Directory);
+
+            // delete dest to break the symlink
+            Directory.Delete(dest);
+
+            Delete.Perform(source);
+
+            Assert.IsFalse(Directory.Exists(source));
+        }
+
+        [TestMethod]
+        public void PerformWorksWithSymbolicFileLink()
+        {
+            if (!WinApiService.IsUserAnAdmin())
+            {
+                Assert.Inconclusive("This test only works on administrative privileges.");
+                return;
+            }
+
+            string source = Util.GetTempPath();
+            string dest = Directory.CreateDirectory(Util.GetTempPath()).FullName;
+
+            WinApiService.CreateSymbolicLink(source, dest, WinApiService.SYMBOLIC_LINK_FLAG.File);
+
+            // delete dest to break the symlink
+            Directory.Delete(dest);
+
+            Delete.Perform(source);
+
+            Assert.IsFalse(File.Exists(source));
         }
 
         [TestMethod]